feat(frame-pacing): support fractional frame generation multipliers
The multiplier becomes a float, so non-integer values like 1.5 or 1.75
work. Each cycle a debt accumulator turns it into an integer generated-
frame count (debt += R-1; frames = floor(debt); debt -= frames), so 1.5
alternates 0 and 1 generated frames for a 0.5 average.

scheduleFrames now takes the per-cycle count: it rewrites each frame's
phase to (i+1)/(frames+1) through a host-visible constant-buffer update
(race-free behind the existing cmdbufFence) and signals the completion
fence from the pre-pass when zero frames are generated. The generator
allocates ceil(R)-1 destination images and fetches them by per-cycle
index. R=2.0 reduces exactly to the prior fixed path.

Verified on 007 First Light (Proton/SLR4, RX 9070 XT): R=2.0 gives
170 -> 340 fps, R=1.5 gives 219 -> 329 fps.

+++ lsfg-vk-common/include/lsfg-vk-common/configuration/config.hpp
@@ -34,8 +34,8 @@ namespace ls {
         std::vector<std::string> active_in;
         /// gpu to use (in case of multiple)
         std::optional<std::string> gpu;
-        /// multiplier for frame generation
-        size_t multiplier{2};
+        /// frame generation multiplier; fractional values are allowed (e.g. 1.5)
+        float multiplier{2.0F};
         /// non-inverted flow scale
         float flow_scale{1.00F};
         /// use performance mode
